+++ core/data.go
@@ -3,8 +3,20 @@ package core
 type RunStatus string
 
 const (
-	RunStatusRunning   RunStatus = "Running"
-	RunStatusSkipped   RunStatus = "Skipped"
-	RunStatusSucceeded RunStatus = "Succeeded"
-	RunStatusFailed    RunStatus = "Failed"
+	RunStatusRunning    RunStatus = "Running"
+	RunStatusSkipped    RunStatus = "Skipped"
+	RunStatusSucceeded  RunStatus = "Succeeded"
+	RunStatusFailed     RunStatus = "Failed"
+	RunStatusTerminated RunStatus = "Terminated"
 )
+
+type RunShow struct {
+	ID            int64
+	JobName       string
+	StartTime     string
+	EndTime       string
+	LogFile       string
+	SystemLogFile string
+	Status        string
+	Duration      string
+}