+++ core/format.go
@@ -1,6 +1,8 @@
 package core
 
-import "time"
+import (
+	"time"
+)
 
 func FormatStatus(status RunStatus) string {
 	switch status {
@@ -9,16 +11,21 @@ func FormatStatus(status RunStatus) string {
 	case RunStatusFailed:
 		return "❌ " + string(status)
 	case RunStatusRunning:
-		return "⚙️ " + string(status)
+		return "🚀 " + string(status)
 	case RunStatusSkipped:
 		return "⚠️ " + string(status)
+	case RunStatusTerminated:
+		return "💥 " + string(status)
 	}
 	return string(status)
 }
 
-func FormatTime(timeValue time.Time) string {
+func FormatTime(timeValue time.Time, useLocalTime bool) string {
 	if timeValue.IsZero() {
 		return ""
+	}
+	if useLocalTime {
+		return timeValue.In(time.Local).String()
 	}
 	return timeValue.String()
 }