+++ notify/notify_test.go
@@ -9,7 +9,7 @@ import (
 func Test_getNotifyText(t *testing.T) {
 	data := []struct {
 		name             string
-		cronName         string
+		jobName          string
 		runId            int64
 		runStatus        core.RunStatus
 		logFile          string
@@ -30,17 +30,16 @@ Log: ` + "`/file/path`"},
 Log: ` + "`/file/path`"},
 		{"notify-log-content", "test-3", 34, core.RunStatusSucceeded, "testdata/example.log", true, "",
 			`*test-3*: run 34 - ✅ Succeeded
-Log: ` + "`testdata/example.log`" + "\nLog Content:\n" + "```" + "\nLine one of log\nLine two of log\n```"},
-		{"notify-log-content-file-does-not-exist", "test-3", 34, core.RunStatusSucceeded, "testdata/notreal.log", true, "",
-			`*test-3*: run 34 - ✅ Succeeded
-Log: ` + "`testdata/notreal.log`"},
+Log:
+` + "```" + "\nLine one of log\nLine two of log\n```"},
+		{"notify-log-content-file-does-not-exist", "test-3", 34, core.RunStatusSucceeded, "testdata/notreal.log", true, "", "*test-3*: run 34 - ✅ Succeeded"},
 	}
 
 	for _, d := range data {
 		t.Run(d.name, func(t *testing.T) {
 			notifyStr := getNotifyText(
 				RunNotifyInfo{
-					Name:             d.cronName,
+					Name:             d.jobName,
 					NotifyLogContent: d.notifyLogContent,
 					Id:               d.runId,
 					Status:           d.runStatus,